Fix type

+++ apis/database/api.go
@@ -86,7 +86,7 @@ func (api *API) GetBlockHeader(blockNum uint32) (*BlockHeader, error) {
 }
 
 // GetBlock return a block by the given block number
-func (api *API) GetBlock(blockNum int64) (*Block, error) {
+func (api *API) GetBlock(blockNum uint32) (*Block, error) {
 	var resp Block
 	err := api.call("get_block", []interface{}{blockNum}, &resp)
 	return &resp, err
